+++ app/versiondb.go
@@ -9,7 +9,7 @@ import (
 	"github.com/crypto-org-chain/cronos-store/versiondb"
 	"github.com/crypto-org-chain/cronos-store/versiondb/tsrocksdb"
 
-	storetypes "cosmossdk.io/store/types"
+	storetypes "github.com/cosmos/cosmos-sdk/store/v2/types"
 )
 
 func (app *App) setupVersionDB(