+++ cmd/cronosd/dbmigrate/patch.go
@@ -14,7 +14,7 @@ import (
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/gogoproto/proto"
 
-	"cosmossdk.io/log"
+	"cosmossdk.io/log/v2"
 )
 
 const (